# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ChatSite AI API
  version: 1.0.0
  description: Multi-tenant RAG chatbot SaaS — all routes prefixed with /api/v1
servers:
  - url: https://api.chatsite.ai/api/v1
  - url: http://localhost:8000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          type: object
        error:
          nullable: true
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    UserRegister:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        full_name: { type: string }
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 900 }
...
    CrawlStart:
      type: object
      required: [website_id]
      properties:
        website_id: { type: string, format: uuid }
        force_recrawl: { type: boolean, default: false }
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRegister' }
      responses:
        '201': { description: Created }
  /auth/login:
    post:
      summary: Email + password login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserLogin' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
...
  /workspaces:
    get:
      summary: List workspaces for current user
      tags: [Workspaces]
      responses: { '200': { description: OK } }
    post:
      summary: Create a workspace
      tags: [Workspaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, subdomain]
              properties:
                name: { type: string }
                subdomain: { type: string }
                plan: { type: string }
      responses: { '201': { description: Created } }
  /workspaces/{id}:
    get:
      tags: [Workspaces]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses: { '200': { description: OK } }
    put:
      tags: [Workspaces]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses: { '200': { description: OK } }
  /websites:
    get:
      tags: [Websites]
      responses: { '200': { description: OK } }
    post:
      tags: [Websites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, name]
              properties:
                url: { type: string, format: uri }
                name: { type: string }
                crawl_config: { type: object, additionalProperties: true }
      responses: { '201': { description: Created } }
  /websites/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:    { tags: [Websites], responses: { '200': { description: OK } } }
    put:    { tags: [Websites], responses: { '200': { description: OK } } }
    delete: { tags: [Websites], responses: { '204': { description: No content } } }
  /crawl/start:
    post:
      tags: [Crawl]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CrawlStart' }
      responses: { '202': { description: Accepted } }
  /crawl/stop:
    post:
      tags: [Crawl]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [website_id]
              properties:
                website_id: { type: string, format: uuid }
      responses: { '202': { description: Accepted } }
  /crawl/reindex:
    post:
      tags: [Crawl]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [website_id]
              properties:
                website_id: { type: string, format: uuid }
      responses: { '202': { description: Accepted } }
  /crawl/status/{website_id}:
    get:
      tags: [Crawl]
      parameters:
        - in: path
          name: website_id
          required: true
          schema: { type: string, format: uuid }
      responses: { '200': { description: OK } }
...